Breakdown websockets into files per core type
