+++ app/pkg/runme/stream/processor.go
@@ -1,0 +1,100 @@
+package stream
+
+import (
+	"context"
+	"io"
+
+	"github.com/jlewi/cloud-assistant/app/pkg/logs"
+	"github.com/jlewi/cloud-assistant/app/pkg/runme"
+	"github.com/pkg/errors"
+	v2 "github.com/runmedev/runme/v3/api/gen/proto/go/runme/runner/v2"
+	"google.golang.org/grpc/metadata"
+)
+
+// RunmeProcessor handles the execution requests and responses for a run.
+type RunmeProcessor struct {
+	Ctx              context.Context
+	RunID            string
+	ExecuteRequests  chan *v2.ExecuteRequest
+	ExecuteResponses chan *v2.ExecuteResponse
+	// StopReading is used to signal to the readMessages goroutine that it should stop reading messages
+	StopReading chan bool
+
+	Runner *runme.Runner
+}
+
+func NewRunmeProcessor(ctx context.Context, runID string) *RunmeProcessor {
+	p := &RunmeProcessor{
+		Ctx:   ctx,
+		RunID: runID,
+		// Create a channel to buffer requests
+		ExecuteRequests:  make(chan *v2.ExecuteRequest, 100),
+		ExecuteResponses: make(chan *v2.ExecuteResponse, 100),
+		StopReading:      make(chan bool, 1),
+	}
+
+	return p
+}
+
+func (p *RunmeProcessor) close() {
+	// Close the requests channel to signal to the Runme that no more requests are expected
+	close(p.ExecuteRequests)
+	// Close the responses channel to signal to the Runme that no more responses are expected
+	close(p.ExecuteResponses)
+}
+
+func (p *RunmeProcessor) Recv() (*v2.ExecuteRequest, error) {
+	log := logs.FromContextWithTrace(p.Ctx)
+
+	req, ok := <-p.ExecuteRequests
+	if !ok {
+		log.Info("Channel closed", "runID", p.RunID)
+		// We return io.EOF to indicate the stream is closed by the client per the grpc Bidi spec.
+		return nil, io.EOF
+	}
+	return req, nil
+}
+
+// Send sends a response message to the client.  The server handler may
+// call Send multiple times to send multiple messages to the client.  An
+// error is returned if the stream was terminated unexpectedly, and the
+// handler method should return, as the stream is no longer usable.
+func (p *RunmeProcessor) Send(res *v2.ExecuteResponse) error {
+	p.ExecuteResponses <- res
+	return nil
+}
+
+func (p *RunmeProcessor) SetHeader(md metadata.MD) error {
+	log := logs.FromContextWithTrace(p.Ctx)
+	log.Info("Set called", "md", md, "runID", p.RunID)
+	return nil
+}
+
+func (p *RunmeProcessor) SendHeader(md metadata.MD) error {
+	log := logs.FromContextWithTrace(p.Ctx)
+	log.Info("SendHeader called", "md", md, "runID", p.RunID)
+	return nil
+}
+
+func (p *RunmeProcessor) SetTrailer(md metadata.MD) {
+	log := logs.FromContextWithTrace(p.Ctx)
+	log.Info("SetTrailer called", "md", md, "runID", p.RunID)
+}
+
+func (p *RunmeProcessor) Context() context.Context {
+	return p.Ctx
+}
+
+func (p *RunmeProcessor) SendMsg(m any) error {
+	err := errors.New("SendMsg is not implemented")
+	log := logs.FromContextWithTrace(p.Ctx)
+	log.Error(err, "SendMsg is not implemented")
+	return err
+}
+
+func (p *RunmeProcessor) RecvMsg(m any) error {
+	err := errors.New("RecvMsg is not implemented")
+	log := logs.FromContextWithTrace(p.Ctx)
+	log.Error(err, "RecvMsg is not implemented")
+	return err
+}
